// shared/types/house.type.ts
export type HouseType = {
    id: number;
	nickname: string;
	address: string;
	address_number:  string;
	address_complement:  string;
	address_neighborhood:  string;
	address_city:  string;
	address_state:  string;
	address_postal_code:  string;
	created_at:  string;
};

export type HouseSelectType = {
    id: number;
	nickname: string;
};

// export const HouseOrderByList = ['Apelido', 'Endereço', ]